requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pytest==7.4.4
pytest-html==4.1.1
requests==2.31.0
pydantic==2.6.0
pytest-xfist==3.0.0
pytest-cov==4.1.0
pytest-timeout==2.2.0
pytest-parallel==1.5.0
python-dotenv==1.0.0
jsonschema==4.21.1